Start scene added
